+++ examples/mvvmfx-contacts/src/main/java/de/saxsys/mvvmfx/contacts/ui/validators/PhoneValidator.java
@@ -1,7 +1,7 @@
 package de.saxsys.mvvmfx.contacts.ui.validators;
 
 import de.saxsys.mvvmfx.utils.validation.ValidationMessage;
-import de.saxsys.mvvmfx.utils.validation.ObservableRuleBasedValidator;
+import de.saxsys.mvvmfx.utils.validation.validators.ObservableRuleBasedValidator;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
 import javafx.beans.value.ObservableValue;
