+++ mvvmfx/src/main/java/de/saxsys/mvvmfx/utils/validation/validators/FunctionBasedValidator.java
@@ -1,5 +1,8 @@
-package de.saxsys.mvvmfx.utils.validation;
+package de.saxsys.mvvmfx.utils.validation.validators;
 
+import de.saxsys.mvvmfx.utils.validation.ValidationMessage;
+import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
+import de.saxsys.mvvmfx.utils.validation.Validator;
 import javafx.beans.value.ObservableValue;
 
 import java.util.Optional;
