+++ mvvmfx/src/main/java/de/saxsys/mvvmfx/utils/validation/validators/ObservableRuleBasedValidator.java
@@ -1,5 +1,8 @@
-package de.saxsys.mvvmfx.utils.validation;
+package de.saxsys.mvvmfx.utils.validation.validators;
 
+import de.saxsys.mvvmfx.utils.validation.ValidationMessage;
+import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
+import de.saxsys.mvvmfx.utils.validation.Validator;
 import javafx.beans.value.ObservableValue;
 
 import java.util.ArrayList;
@@ -43,7 +46,7 @@ public class ObservableRuleBasedValidator implements Validator{
      * the rule is violated. In this case the given message object will be added to the status
      * of this validator.
      * 
-     * There are some predefined rules for common use cases in the {@link Rules} class that can be used. 
+     * There are some predefined rules for common use cases in the {@link ObservableRules} class that can be used.
      * 
      * @param rule
      * @param message
