+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/utils/validation/validators/RulesTest.java
@@ -1,5 +1,6 @@
-package de.saxsys.mvvmfx.utils.validation;
+package de.saxsys.mvvmfx.utils.validation.validators;
 
+import de.saxsys.mvvmfx.utils.validation.validators.ObservableRules;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.beans.value.ObservableBooleanValue;
@@ -19,7 +20,7 @@ public class RulesTest {
 	public void testNotNullOrEmpty() {
 		StringProperty value = new SimpleStringProperty();
 
-		final ObservableBooleanValue rule = Rules.notEmpty(value);
+		final ObservableBooleanValue rule = ObservableRules.notEmpty(value);
 		
 		assertThat(rule.get()).isFalse();
 
@@ -47,7 +48,7 @@ public class RulesTest {
 	public void testMatches() {
 		StringProperty value = new SimpleStringProperty();
 
-		final ObservableBooleanValue rule = Rules.matches(value, Pattern.compile("[0-9]"));
+		final ObservableBooleanValue rule = ObservableRules.matches(value, Pattern.compile("[0-9]"));
 		
 		assertThat(rule.get()).isFalse();
